git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@785 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ src/serialization/gmetastorage_json.cpp
@@ -235,7 +235,7 @@ const char * GJsonNodeNameTracker::makeName(const char * name, bool addNameCount
 
 	if(count > 0) {
 		stringstream stream;
-		stream << name << count;
+		stream << name << (unsigned int)count;
 		this->nameBuffer = stream.str();
 		result = this->nameBuffer.c_str();
 	}
